fix: default schema locales overwrite gcms-locales header
For instance the pages() query of the Hygraph systeem has the following
schema defs:

locales: [Locale!]! = [en]

When running GraphQL mesh with multiple upstream services, the schemas
are being stitched and schema defaults are added to the query before sending
it upstream.

Meaning the base query

query {
  pages(where: { url: "service"}) {
    title
    __typename
    metaTitle
    author
  }
}

Is quietly adjusted to

query {
  pages(where: { url: "service"}, locales: "en") {
    title
    __typename
    metaTitle
    author
  }
}

Hygraph gives a higher priority to the locales being passed as parameter
than the locales being passed through the gcms-locales header.

This resulted in all query results being in English.

To circumvent this for now, we strip all locales information to prevent
the mesh from sending the schema defaults upstream

    transforms:
      - filterSchema:
          filters:
            - 'Query.*.!locales'