Footer components should be rendered conditionally
If, for example, storeSwitcher is not passed, the corresponding container shoudn’t be rendered either. This prevents unnecessary HTML.

—

Gaps could be created with the current grid-template when removing an item. The grid could be rewritten to respond to adding or removing items, but that seems a bit to complex for a demo template. Developers can also easily re-assign a component to another grid-area.