Use 'cache-first’ fetch mode for getting 'total_quantity' data (GCOM-1241)
We need to create a cache entry to read it on the next reload. This change helps us store the 'total_quantity' data in cache, so we can read it again when needed.

Now, when a guest adds products to their cart and refreshes the page, the cart will still show what's in it. It won't be empty.