refactor: Destructure InputProps to ensure they are passed only to relevant components.
- This resolves the following error when NumberFieldElement has InputProps:

`<NumberFieldElement InputProps={{ disableUnderline: true }} />`

```
Warning: React does not recognize the `disableUnderline` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `disableunderline` instead. If you accidentally passed it from a parent component, remove it from the DOM element. Error Component Stack
```